Add copied_files option
The copied_files option allows templates to specify file formats that
should not be processed with EEx, but should instead simply be copied to
the new directory.

Add messages when files that are intended to be processed with EEx
encounter some error. Instead of erroring and stopping the entire
template generation with an ugly error, report that files which
encounter errors are being ignored in red, and display a reason.